// src/queue/enums/queue-name.enum.ts
export enum QueueName {
  EMAIL_GOOGLE_SYNC_INIT = 'email-google-sync-init',
  EMAIL_GOOGLE_SYNC_CHUNK = 'email-google-sync-chunk',
  EMAIL_GOOGLE_SYNC_PAGE = 'email-google-sync-page',
  CALENDAR_GOOGLE_SYNC_INIT = 'calendar-google-sync-init',
  CALENDAR_GOOGLE_SYNC_CHUNK = 'calendar-google-sync-chunk',
  CALENDAR_GOOGLE_SYNC_PAGE = 'calendar-google-sync-page',
  EMAIL_MICROSOFT_SYNC_INIT = 'email-microsoft-sync-init',
  EMAIL_MICROSOFT_SYNC_CHUNK = 'email-microsoft-sync-chunk',
  EMAIL_MICROSOFT_SYNC_PAGE = 'email-microsoft-sync-page',
  CALENDAR_MICROSOFT_SYNC_INIT = 'calendar-microsoft-sync-init',
  CALENDAR_MICROSOFT_SYNC_CHUNK = 'calendar-microsoft-sync-chunk',
  CALENDAR_MICROSOFT_SYNC_PAGE = 'calendar-microsoft-sync-page',
}